Document changes for startState and extraConstants
Documents the newest model changes in babel-plugin-automaton regarding
startStates which are no longer only strings but can be transition
functions now.

In turn the data extracted for them is not a string anymore but the
actual compact code. (string literal or function expression)

extraConstants allows defining arbitrary additional constants and
functions in the process module.