Implement automaton graphql service
So far, every DomainQL-project home-brew their own graphql service based
on fetch, this now standardizes a client-side GraphQL-service for
automaton. Most of it is not really automaton specific, but there were
enough places where it involves minor automaton specific conventions
that would have made the same implementation on a DomainQL level much
more awkward. Other DomainQL-users may find inspiration in this.

We need to be able to do a WireFormat conversion both when receiving
data from the GraphQL service as well as when sending data as input to a
GraphQL service.

Converting results
==================

To convert results, we look use the fact that the name of the query
method is reflected in the response data so that we can look up the
GraphQL method in the schema data and thus find out which GraphQL type
is returned from that method.

This in turn allows us to both do all nescessary conversions of embedded
scalar values as well as instantiated the appropriate mobx model classes
where such classes are defined in the local application below

    src/main/js/apps/<app-name>/domain

Converting Input Data
=====================

To be able to convert the input data, we use the graphql package's
parser to parse the query to then get to the inputs defined for the
client-side query.

We transform the former QueryDeclaration only used for injection into a
general client-side abstraction for GraphQL queries/mutations that will
cache the evaluated input types to be able to do this once.

If we receive a string query, we convert that live within the graphql
method.

Application authors are recommened to use a constant to store their
GraphQL queries to then call `query.execute(params)` to execute it.

Adds a simple renderer to prettyish-print GraphQL errors to the
console.

Adds tests for the query parsing