Fine-tune automaton button
Make the automaton button work with the transition mechanism and also
observe values to react to value changes within a form appropriately.

Adds support for  confirmations on the transition.

If the property `confirmation` is a function (receiving the current
context) that returns a truthy string, that string will be displayed
as confirmation message.

If the confirmation methods returns something falsy, there will be
no confirmation and the transition will continue.

If the the user denies the confirmation, the transition will be
aborted (before the view model is created)

Disabling
=========

If the disabled prop is given as function, that function will be called
first with the formConfig.

If no such prop exists or if the prop returns true, a transition button
will be disabled if the transition is not discarding and the formConfig
has errors.

Transition Context
==================

If the context is set by prop, that context will always be used.
Otherwise we use the current form state as context for the transition
triggered by the button.