Experimental: Refactor transition / history
This commit changes the way we're connecting the current process state
to the browser history and what we store in the navigation history
entries.

So far, every transition would create a new browser history state,
mostly without the URL changing.

The browser history state would also only snapshot the current state
name and restore that later without changing important process scope
members representing current object selections. So after going to the
detail view of multiple objects in "crud-test" for example, the history
would then jump back and forth the list and detail view, but keep the
`currentFoo` selection thus always showing the object selected last.

So apparently we need some scope members that are in fact snapshotted
with the browser history state and restored when the user navigates back
to that history entry. Spring Webflow would do that for all process scope
variables, but I think that doing so only selectively is actually the
better solution and memory-saving, too.

Versioning
==========

So we give the Process a versioningStrategy function property that
defines which properties are to be versioned and which are not.
`process.versioningStrategy(name)` must return true if the property with
the given name is to be versioned, that is copied to the browser history
state and restored from there in case of the user navigating back.

For the default implementation we use all properties whose name starts with
"current".

Recording
=========

Since we want to limit the history creation for meaningful states we
have to deal with a transition either being recorded in the browser
history or not. The rules according to which seem still a bit fuzzy, but
my current thinking and implementation goes as far:

Transitions that change view states obviously are part of the history.
Transitions that stay within the same view should most often *not* be
recorded, unless they change for example our `currentFoo` object for the
current selection. Navigation from one form object to another is
obviously an important detail.

At first I added a "isRecorded" field to transition which controlled the
recording and had to flipped manually by the user.

But it seems that our example of a versioned field changing
necessitating a transition recording might always be the reason.

So we make it a rule.

Transition Scope ViewModel
==========================

Early on I wrote the transition mechanic so that the scope within the
transition was in fact a view model of that scope allowing all changes
within that transition to be applied afterwards and also be canceled.

But this worked never quite right, because it only caught access via
`process.scope` while the more natural and clearly intended use of
`scope` did *not* catch those changes.

To implement this we'd need another complicated proxy to proxy `scope`
from the beginning and then hot swap etc pp.

In the end, all that does not seem worth it.

Especially since the transition action being a mobx action leads
transitional behavior anyway and `transition.cancel()` was more of a
cute idea than having a strong use-case.

We now keep the scope and use mobx observe() to detect changes to
versioned fields.

We keep transition.isRecorded around as a way to force recording in case
I overlooked something.