Make t.back() and browser navigation the same
So given the intrinsic connection of the `transition.back(n)` function
and the internal browser history we have decided that using that
function is identical to the user using browser back.

We change executeTransition to resolve to the transition instance as it
is after the transition execution.

The back function now just locates the index of the history entry
resulting from the given numeric or function parameter.

After the transition is executed, we use that navigate our history
object to that index, making the functionality technically virtual
identical except for the source of the history navigation.

We make sure we can forbid transition recording to allow all possibility
for application authors.

Refactors the creation of our mobx action functions into a
`prepareMobXAction` function.