Always use executeTransition()
Transitions without action functions would be short-cutted before
entering executeTransition() which was in fact more of a
executeTransitionAction().

Routing through executeTransition() in any case harmonizes the
post-transition handling for simple transitions and simple start
transitions and fixes erroneous behavior for those

We streamline executeTransition() a bit and make it handle a null action
correctly.