Add ScrollTracker component
Add a new component from our test project.

It needs to be fed with a current formConfig and stabilizes the
currently focused form element during updates so that changes in form
errors don't make the input jump, but the view is scrolled perfectly to
line up the input to where it was before.

Class component because we need getSnapshotBeforeUpdate.