Add System Types
So far we have been able to instantiate user-defined types when
doing the json -> js conversion (see WireFormat.js).

We now add the ability to do so programmatically via startup and also
based on a generic Java type.

We add an empty registerSystemTypes where we will add system-behaviors
implemented as client-side domain classes