Implement lookupType/unwrapAll helpers
While we are able to resolve types for e.g. <Field name="foos.1.name" />,
we need an alternate way of resolving types the way GraphQL handles
these things in queries.

For example for a query like

    query iQueryNode($config: QueryConfigInput)
    {
        iQueryNode(config: $config)
        {
            rows{
                id
                name
            }
            rowCount
        }
    }

the `rows` just the the `foos` in the first example is assumed to be a
GraphQL list, but in the first example, we need to define the actual
index within the list, while for the second, when referencing a field
it would naturally be `iQueryNode.rows.name` without the index,
effectively ignoring the list.

This commit adds two utility modules. `unwrapAll` unwraps a GraphQL type
reference, without all NON_NULL and LIST wrappers.

lookupType(complexTypeName, path) can resolve the type of a field within
a GraphQL query.