Implement Filter DSL mirroring JOOQ
This commit provides the Javascript side implementation of the
`Condition` scalar.

It copies the Java condition API of JOOQ as much as possible but with
another way of referencing fields and constant values as necessary for
the interactive query implementation.

Let's consider an (incomplete) interative query like

    query iQueryFoo($config: QueryConfigInput!)
    {
        iQueryFoo(config: $config)
        {
            rows{
                id
                name
                num
                description
                created
                type
                flag
                owner{
                    id
                    login
                }

            }
        }
    }

seletion on the `rows` field and sub-fields.

For the purpose of the interactive query, this query has the following
fields

    * id
    * name
    * num
    * description
    * created
    * type
    * flag
    * owner.id
    * owner.login

A list of all scalar values within the rows field. Note how we're
following the owner field (to `AppUser`) and how the owners fields
have this dot notation (e.g. "owner.login")

```js
    import { FilterDSL } from "@quinscape/automaton-js"
    // deconstruct FilterDSL methods
    const { field, value } = FilterDSL;

    const filter = field("name").equalIgnoreCase(
            value("String", "abc")
        );

```

The comparisons available at this point are all the non-deprecated JOOQ
comparisons. The DSL also supports the non-deprecated operations on
fields.

```js
    import { FilterDSL } from "@quinscape/automaton-js"
    // deconstruct FilterDSL methods
    const {  field, value } = FilterDSL;

    const filter = field("numA").add(
                field("numB")
            ).gt(
                value("Int", 100)
            )

```

Constant values have to be wrapped in a value() node defining the scalar
type of the value and the value itself. This is necessary to ensure
proper conversion of all scalar types.

Conditions can be combined with logical operators, both in the field
syntax

```js
    import { FilterDSL } from "@quinscape/automaton-js"
    // deconstruct FilterDSL methods
    const { field, value } = FilterDSL;

    const filter = field("name").eq(
        value("String", "abc")
    ).and(
        field("num").eq(
            value("Int", 12)
        )
    );
```

as well as a with independent logical conditions

```js
    import { FilterDSL } from "@quinscape/automaton-js"
    // deconstruct FilterDSL methods
    const { field, value, and, or, not } = FilterDSL;

    const filter = and(
        field("name").eq(
            value("String", "abc")
        ),
        or(
            field("num").eq(
                value("Int", 12)
            ),
            not(
                field("flag").isTrue(),
            )
        )
    );

```