useObservableInput hook
This is a somewhat experimental hook to connect Mobx observables to the
react hook input mechanism.

In some cases, we have to deal with both Mobx observables and hook based
state in the same component. I don't think it makes very much sense to
move absolutely *everything* over to Mobx, even the tiniest bit of
compoment states. But without doing so we will always have situations
where both worlds collide.

So for the new IQuery grid for example, I have a memoized `columns`
structure that contains everything the component needs to know about its
columns and which is a mix of observable `ColumnConfig` and domain type
resolution and some column props mixed in.

With configurable columns we have the very rare occasion that the
`columnConfig` actually changes for a grid and since the definition of
that behavior is local and outside of the iQuery mechanism, we have no
idea when.

We just know that the columnConfig object changes in content.

Using it directly as `useMemo()` input does not work because the
observables are mutable and the instance does not actually change.

So what this hook does is to setup a specialized reaction in such a
case.

    const input = useObservableInput( () => myObservable.prop )

the hook sets up an reaction for the given expression and if that
changes increments the input value which in turn can be used as hook
input in place of the original observable.

Just make sure your expression function touches all relevant observables
when you have a larger structure.