DataGrid.Column: Add renderFilter
Adds a new render function prop to the filter row to allow the use to
define custom filter elements.

The render function takes a function of the form

    (fieldName, fieldType, label, index) => (<input ... />)

and returns a react element for the given parameters

    * fieldName - generate name/path within the filter row object
    * fieldType - scalar type of the column
    * label     - (hidden) label for the element
    * index     - value index for the column (for multiple inputs per
                  column)

The render function will be called once for each value argument of the
filter associated with the column.