Add FKSelector widget
This commit adds a form widget to update foreign key reference fields
and embedded objects of the same relation.

The widget renders a disabled text-field with a "..." button addon that
opens up a modal dialog with a data grid.

    <FKSelector
        name="quxCId1"
        display={ () => root.quxC1.name }
        onUpdate={
            row => formConfig.root.quxC1 = row
        }
        query={
           new GraphQLQuery(
                // language=GraphQL
                `
                query iQueryQuxA($config: QueryConfigInput!)
                {
                    iQueryQuxA(config: $config)
                    {
                        # ...
                    }
                }`,
                {
                    "config": {
                        "pageSize": 5,
                        "sortFields": ["name"]
                    }
                }
            )

        }
    />

The optional name references the foreign key field directly. If the name
prop is given fk-selector will always update the foreign key value.

The display property is a render function rendering the current value.
You don't really need it if the foreign key value is a natural and not
an id.

`onUpdate` is called when the user selects a new row value (might be
null for nullable fields). Without onUpdate, the fk-selector tries to
update the embedded objects of the same type within the in-memory output
type (coming from a previous query). If onUpdate is set, it is the users
responsibility to update all values in it (except for the foreign key value
itself)

query defines an iQuery GraphQL-query that is provides the data
displayed in the modal data grid.