Add static query() function
So far, we've been defining GraphQLQuery objects with their constructor,
but babel-plugin-track-usage does not track constructors (maybe it
should).

For now we evade that with a static method `query(query,variables)`.

we change injection() so that it just uses the first argument if it
already is a GraphQLQuery instance.

This is key for a new feature where we externalize injected queries,
allowing a `inject( Q_QuxMain)` to infer that we mean to inject the
query result of a named query.

Named Queries
=============

Named queries must be in a "queries/" folder. Their name is determined
by the name of the module they are in.

    import { query } from "@quinscape/automaton-js"

    export default query(
        // language=GraphQL
            `query iQueryQuxMain($config: QueryConfigInput!)
        {
            iQueryQuxMain(config: $config)
            {
                rows{
                    id
                    name
                }
            }
        }`,
        {
            "config" : {
                "pageSize" : 5
            }
        }
    )

The static query method uses the same arguments as the GraphQLQuery
constructor.

The server-side process-injection detects the invocation of injection()
with an identifier and looks up that identifier among the named modules
completing the indirected injection.