DevUtil: pickSchemaTypes
This is an developement utility function to produce minimized GraphQL
schemas from a larger project schema.

For our purposes, we always work with the JSON variant of a GraphQL
schema i.e. the result of the introspection query.

This commit adds a function `pickSchemaTypes(rawSchema, typeNames)` that
is given such a raw JSON GraphQL schema and an array with some types
within that schema.

The method will then return a minimized Schema that only contains the
types reachable from the given types.