Startup: Add indirection for injections logging
Injections with their giant query string keys keep being cumbersome to
use in the browser console.

We already added a separate "INJECTED VALUES" log, but the injections of
course still appear in the initial data block.

We hide the injections by using Object spread to introduce another map
with "..." pointing to the actual injections.

This way  the browser only renders the "..." key but not the injection
keys at first but the user can still open and inspect the object if they
want.