Refactor FKSelector
This commit changes the preexisting <FKSelector/> component to allow
textual-input to change the foreign key target.

The input value is stored as local state. The component does a debounced
validation of the input by querying the GraphQL query with a filter
resulting from the input.

The `validateInput` prop controls this behavior. It can be either a
String or a function returning a filter expression

    <FKSelector
        label="quxD"
        display="quxD.name"
        validateInput="name"
        query={ Q_QuxD }
    />

The "name" value of the prop is converted into a `field("name").eq(val)`
filter expression.

    <FKSelector
        label="quxD2"
        display="quxD.name"
        validateInput={
            val => or(
                field("id")
                    .startsWith(
                        value(
                            "String",
                            val
                        )
                    ),
                field("name")
                    .startsWith(
                        value(
                            "String",
                            val
                        )
                    )
            )
        }
        query={ Q_QuxD }
    />

The input is valid when the query results in exactly one result. In this
case, the result is used as selected foreign key object.

This also adds the ability to add additional addons to the <FKSelector/>
to render multiple object fields for a single foreign key.

In the second example above, e.g. we can add the `id` field as Addon.

    <FKSelector
        label="quxD2"
        display="quxD.name"
        validateInput={
            val => or(
                field("id")
                    .startsWith(
                        value(
                            "String",
                            val
                        )
                    ),
                field("name")
                    .startsWith(
                        value(
                            "String",
                            val
                        )
                    )
            )
        }
        query={ Q_QuxD }
    >
        <Addon placement={ Addon.RIGHT} text={ true } className="text-monospace">
            {
                () => (currentQux && currentQux.quxD && currentQux.quxD.id) || GlobalConfig.none()
            }
        </Addon>

    </FKSelector>

The JSX expression containing an arrow function can be used to evade
observable problems with the addons.

This adds half a test for the new input feature in that it adds two
input examples to the existing test. It does not yet test the update
behavior