FilterDSL: Support Values nodes, IN operator
So far, we've been missing the IN operator which we add now. It uses a
collection of scalar values as input.

For this purpose we add a new node type "Values" which is just like the
"Value" type, but it contains an array of scalar values of the same
type.

    query.execute({
        config: {
            condition:
                field("id").in(
                    values("Int", [1, 2, 3])
                ),
            pageSize: 0
        }
    })