Transition: Fix mobx action rep
For some reason we only considered the transition name as key for mobx
actions we create for transition functions.

Since the transition names are only unique per containing state, this is
clearly wrong and leads to all "save" transitions sharing the same
action.

Adding the state prefix solves that issue.