Implement <AssociationSelector/>
Implements the new <AssociationSelector/> selector component that
renders a bootstrap 4 list group for every entity the current entity is
connected to via the associative entity.

Here we see an example of an associative entity "BazLink" as
many-to-many connection between the associated domain types "Baz" and
"BazValue"

    "Associated Entity A (see BazLink)"
    type Baz {
        "Many-to-many objects from 'baz_link.baz_id'"
        bazLinks: [BazLink]!
        "DB column 'id'"
        id: String!
        "DB column 'name'"
        name: String!
    }

    "Associative entity m-to-n connecting Baz and BazValue"
    type BazLink {
        "Target of 'baz_id'"
        baz: Baz!
        "DB foreign key column 'baz_id'"
        bazId: String!
        "DB column 'id'"
        id: String!
        "Target of 'value_id'"
        value: BazValue!
        "DB foreign key column 'value_id'"
        valueId: String!
    }

    "Associated Entity B (see BazLink)"
    type BazValue {
        "Many-to-many objects from 'baz_link.value_id'"
        bazLinks: [BazLink]!
        "DB column 'id'"
        id: String!
        "DB column 'name'"
        name: String!
    }

From the "Baz" size, the <AssociationSelector/> is used like this

    <AssociationSelector
        name="bazLinks"
        label="Associated Values"
        display="value.name"
        value="value.id"
        helpText="Select associated baz values"
        query={
            Q_BazValueList
        }

    />

The "Baz" object the form is based on must select all its bazLinks
objects including the embedded object of the other side ("value" from
the "Baz" side)

The Q_BazValueList query queries "BazValue" objects without the embedded
bazLinks.

As it is now, the InteractiveQuery mechanism is not able to follow the
many-to-many connections (we might add that later)

At this point, you need a decidated non-iQuery end-point to fetch "Baz"
object including the bazLinks.value sub-objects.