Fix endless websocket connection attempts
When the server restarts and the connection id is not known to the
server, we have no associated session and security context for the
websocket connection and can't really allow a connection.

The server now sends a special error return code 4100 for this
situation.

Detects the return code and notifies the user with an alert that they
should reload the page and then exits the reconnection loop.