Transition: Introduce button attribute
In general we can only influence transition behavior with the one
context object and by generally changing scoped values.

For the most simple use case of having transitions that do exactly
the same but then go to different locations, this is kind of awkward.

We have to introduce a special scoped value, write a custom click
handler and then trigger the transition.

To aid with this use case, we give the transition object a `button`
property that is set when a transition is triggered by automaton-js'
<Button> component in case the button has a name attribute.

So when we have two buttons `<Button name="save" ... />` and
`<Button name="save-and-return" ... />` we can do a conditional
target assignment.

```
    if (t.button === "save-and-return")
    {
        t.target = "ParentView";
    }

```