Implement WorkingSet
This commit contains the client side functionality of the new working
set feature.

The basic idea is that we want to record domain object changes,
creations and deletions before we persist everything in one server
operation.

WorkingSet
==========

class WorkingSet
{
    @action
    addNew(domainObject)
    {
        /// ...
    }

    @action
    addChanges(domainObject)
    {
        /// ...
    }

    @action
    markDeleted(domainObject)
    {
        /// ...
    }

    @action
    revert(domainObject)
    {
        /// ...
    }

    @action
    revertAll()
    {
        /// ...
    }

    lookup(type, id)
    {
        /// ...
    }

    @computed
    get changes()
    {
        /// ...
    }

    @computed
    get newObjects()
    {
        /// ...
    }

    @computed
    get deletions()
    {
        /// ...
    }

    get hasChanges()
    {
        /// ...
    }

    persist(preprocessFn = extractTypeDataFromObjects)
    {
        /// ...
    }
}

The WorkingSet class provides the API to handle these changes. It uses
an internal observable map to store the changes and methods to register
and access the changes or revert them.

The persist() method allows to store all changes.

DataGrid
========

The <DataGrid/> component now accepts a `workingSet` prop to mix-in the
values from a working set to the current objects stored in the database.

New objects appear on the first page of data, changed objects are
replaced with the current state in the working set.

The rendered <tr class="data"> rows receive an additional class for
working set objects ("new-object", "changed-objects", and
"deleted-objects")

<DataGrid.WorkingSetStatus/> is a helper component that renders a status
icon for the current row base on the working set status of the object
row.