WorkingSet: Refactor newObjects
As it was now, a working set would show all new-objects within a
connected data-set and not just the objects of the type the data grid is
showing.

This changes the computer property newObjects to a memoized method
`.newObjects(type)` with an optional type name. If the type name is
given, only objects with this type are returned, if none is given, all
objects are returned.