Remove <Icon/> component
For some reason, we had the identical <Icon/> component both here and in
domainql-form, which does not really make sense.

We can't work without domainql-form, so we have access to the Icon there
anyway.

This commit removes the Icon.js file and replaces all usages with the
domainql-form icon.

Upgrades domainql-form to v3.0.1 for the latest Icon changes.