Refactor Transition.back()
While it still seems good to have "back" concept to simplify a situation
where a view state is transitioned to from multiple places and we just
want to go back to where we came from without having a dedicated
transition for each potential source (plus a scoped variable to track
origin and a mapping ...)

The decision to couple t.back() to the browser history was total
nonsense though. It works on a purely functional level, but it is highly
confusing to the user.

The User edits a domain object, clicks "saves". If the transition
happened with the old type back() however, the user would have to return
to the edit page by clicking "forward" which from their point of view is
at least highly confusing if not outright perceived as an error.

This commit changes the behavior so that t.back() now always creates a new
history entry with the target state.

transition.historyIndex
-----------------------

Setting the `historyIndex` property on a transition is now an inactive
feature. We keep it around for now because there still might be
situations in which we do want a control over browser history.