Refactor Button
First, the kind of convenient `icon` prop generated a lot of issues in
that we couldn't configure the icon (brand icon? regular? light?) but
also that we might want  to have an SVG-icon or PNG icon or whatever. Or
two icons or the text before the icon.

This commit removes the `icon` prop and changes the meaning of the
`text` prop.

We now copy the HTML button in that <Button/> now wraps arbitrary
content.

The `text` property now only controls an additional aria-label property.
This property *must* be set if the <Button/> body does not contain a
textual description.

In development mode the Button actually checks whether there is
printable characters in either body or text prop and throws an error if
that is not the case. Harsh, but this way react provides us with an
actual component stack so we can find the offending button.

If you want an icon within a button now, use either the Icon component
in domainql-form or another icon type.