Implement alternate styles, <StyleSwitcher/>
Client-side implementation of the alternate styles feature.

We add an alternateStyles config variable which we initialize from
the initial data. This requires the AlternateStyleProvider being
used on the server-side. If the provider is not used, alternateStyles
is null.

We add a <StyleSwitcher/> component which allows the user to select
between the alternate styles if present.

StyleSwitcher.getDefaultStylesheet() returns the current default
stylesheet which is either the first style or the user-overridden
style stored as cookie.

StyleSwitcher.setActiveStyleSheet(style) activates the style with the
given name in the current view.

StyleSwitcher.rememberStyle(style) remembers the given style as cookie.

Either use the <StyleSwitcher/> component or write your own using above
static methods and the `config.alternateStyles` info.

AlternateStyles pseudo-GraphQL:
-------------------------------

    {
        currentStyleSheet: String,
        stylesheets: [{ name: String!, uri: String! }]!
    }