DataGrid.Column: Enforce props contract
A <DataGrid.Column/> must either have a single function as child which
receives the current row and renders the cell content or it must have a
name attribute containing a path within the current row to render as a
value.

So far we've been silently rendering "nothing" ("---" by default) when
no name prop was given, but no function child was present.

Now we throw an error in that case.