Add process.isSubProcess
For some purposes it might be useful to have a process that can be
started both as a sub-process and as root process. In general we already
allow that by making `forceSubProcess` optional.

This adds a new getter process.isSubProcess that returns true when the
current process was launched as a sub-process.