Implement accessible Tree widget
In spite our best efforts we were not able to convince the client not to
use a tree as central UI concept. Tree widgets not being optimal from
usability as well as design view points. They however insisted that they
needed that tree, and if we're doing this, we're doing it as best as
possible, following the recommended WAI-ARIA practices
(See https://www.w3.org/TR/wai-aria-practices/#TreeView )

Apart from these accessibility concerns the main design goal was to have
a flexible reusable tree widget that can merge heterogenous domain data
into the hierarchical tree view.

This adds the tree view as a set of composable sub widgets.

Tree.Objects
============

Tree.Objects renders an exisiting iQuery document as tree items.

<h1 id="animal-tree-title" className="sr-only">Animals</h1>
<Tree
    id="animal-tree"
    aria-labelledby="animal-tree-title"
>
    <Tree.Objects
        values={ iQuery }
        render={ (row,isSelected) => row.name }
        actions={
            [
                {
                    label: "Default Action",
                    action: row => ...
                },
                {
                    label: "Action 2",
                    action: row => ...
                }
            ]
        }
    />
</Tree>

Each row object within `iQuery` is rendered by calling the `render`
function with the current row object and the current selection status.

<Tree.Objects/> can have a render function child receiving the current
row object. The render function can render the descendant tree objects
for the row object.

The `actions` prop receives an array of action entries. The action
entries have a `label` and an `action` property. The `action` property
is a function receiving the row object for which the action was
triggered.

The first action is the default action which is triggered when the user
clicks on the name or presses return while the name is focused.

All actions are presented in a context menu which can be opened by
right-clicking on the item name or clicking while pressed ctrl, shift or
alt.

With the keyboard the menu can be opened by either pressing ctrl, shift
or alt while pressing the return key or by tabbing onto a visually
hidden context menu button with opens the same menu. This preserves
discoverability for screen reader users while also enabling quick
access.

Following the WAI-ARIA practices the tree widget is only present twice
within the tab order of the page (current item and context menu button
for that item).

The navigation between items is done with cursor keys as recommended.

Supported are cursor up and down as well as the "home"/"pos 1" key and
the "end" key.

Cursor left can be used to close the current level or jump up to the
current parent if the item has no descendants or it is already closed.

The right cursor opens the current item or jumps to the first child if
the item is already open.

Tree.IndexedObjects
===================

IndexedObjects works very similar to Objects but it arranges the row
values in groups by initial character.

<Tree
    id="animal-tree" aria-labelledby="animal-tree-title">
    <Tree.IndexedObjects
        values={ iQuery }
        index={ indexArray }
        render={(row, iSelected) => row.name }
        renderIndex={letter => (<b>{letter + ":"}</b>)}
        actions={
            [
                {
                    label: "Default Action",
                    action: row => ...
                },
                {
                    label: "Action 2",
                    action: row => ...
                }
            ]
        }
    />
</Tree>

It receives and addition `index` prop which must be provided with an
array of initial characters in sort order.

The optional `renderIndex` render prop is used to render the initial
letter tree group headers. Here we render it bold and add a colon.

Tree.Folder
===========

Finally we have <Tree.Folder/> which is rendered as an initially closed
folder. When the user opens the folder, an iQuery-GraphQL request is
issued and the first result rows are passed onto a render function child
that renders the sub items for the folder.

<Tree.Folder
    render={ () => "Folder Name" }
    query={ <graphql-query> }
    variables={ ... }
>
    {
        iQuery => (
            <Tree.Objects
                values={ iQuery }
                render={ row =>  row.name }
                actions={
                    [
                        {
                            label: "Default Action",
                            action: row => ...
                        },
                        {
                            label: "Action 2",
                            action: row => ...
                        }
                    ]
                }
            />
        )
    }
</Tree.Folder>

We can use <Tree.Objects/> or <Tree.IndexedObjects/> to render the
iQuery result.

The `render` attribute is optional. If it is omitted, the folder will
execute the request on component mount. It will be effectively an
invisible folder from the UI perpective, i.e. it adds its descendants
directly onto the current level without adding a folder level.