Change more selection behavior
The "reselect the same index" works fine when the selection is above or
on the "More" item, but leads to the selection moving when it's below
the "More" and items get inserted before the selection.

Since the components in which the more is handled do not know the
selectionId of the more item, we pass on a "wasSelected" boolean to
the loadMore() function and only do the reselection if we actually need
it because the selection is on the more element.