Implement client-side pub sub
Adds the client-side implementation of the new automaton pubsub feature.

subscribeToTopic
================

    subscribeToTopic(
        "ExampleChannel",   // topic name
        msg => { ... },
        filter
    )

registers the given handler function for a given topic name. The
function will be called for every message published on the channel.

The filter argument is an optional filter to apply to the message
payload. The shape of the payload is channel dependent and basically an
agreement between all participients (server-side services, js code).

The message filter field references must match the shape of the message
payload.

On the other side, we have publish.

    publish(
        "ExampleChannel",
        message
    );

Publishes the given message object by sending it to the server who in
turn publishes it to all subscriptions.

The server-side implementation handles an arbitrary amount of
subscriptions per single physical websocket connection. Each connection
is supplied an numerical id.

This allows each message to be delivered over websocket once and then to
be routed to the right handler functions.