IndexedObjects: Add altText prop
For better a11y it was recommended that we use an additional alt-text
for the initial letter toggle items that have no actions on their own
but instead do that same as the caret button next to it.

We add an altText function prop that produces the alt text for every
index item (letter => altText) and use
`i18n("Toggle Items starting with {0}", letter)` as default.

The tester suggested wrapping both care button and button into one
<div role="button"> but the HTML structure is bad for that as we would
wrap the whole submenu in the button role.

This HTML structure however allows us to have the caret button as big
vertical button target that is easier with the mouse to hit because we
don't realy on the very small caret button itself.

So we just add alt attributes to the index buttons.