Implement domain monitor
Adds the client-side implementation of the domain monitor consisting of
two hooks and a indicator component.

useEntity
=========

    useEntity(domainType, id, version)

Registers the client as currently interacting with a specific entity.
The domain type and the id define the entity. If the Domain monitor is
to be used in concert with the future merging, an entity version is
required, too. Otherwise it is optional.

useDomainMonitor
================

    const monitor = useDomainMonitor(filter, deps)

Creates a domain monitor object that receives all domain activity
updates that match the given filter (most often something like
`field("domainType").eq(value("String", "Foo"))` limiting the updates to
those from a domain type "Foo" ).

If the filter contains values that change over time, the optional deps
array can be used to define the values as inputs dependencies for the
hook.

DomainActivityIndicator
=======================

<DomainActivityIndicator/> display the current activity status of a
single entity. The current implementation renders multiple bootstrap badges
with icons representing ACTIVE, CHANGED and DELETED status with an added
"x <n>" if multiple user are active on an entity.

Example:

    <DomainActivityIndicator
        monitor={ monitor }
        domainType="Foo"
        id={ foo.id }
    />